requirements.txt
apparmor==2.13.5
appdirs==1.4.4
argcomplete==1.8.1
argon2-cffi==18.3.0
arrow==0.15.7
artifacts==20190320
asciitree==0.3.3
asgiref==3.3.1
asn1crypto==1.4.0
async-timeout==3.0.1
asysocks==0.0.2
attrs==20.3.0
autobahn==17.10.1
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
backdoor-factory==0.0.0
bcrypt==3.1.7
beautifulsoup4==4.9.3
binwalk==2.2.0
bleach==3.1.5
blinker==1.4
boltons==19.1.0
bottle==0.12.15
Brotli==1.0.9
cairocffi==0.9.0
capstone==3.0.5
cbor==1.0.0
Cerberus==1.3.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
CherryPy==8.9.1
chirp==0.3.0.dev0
click==7.1.2
click-plugins==1.1.1
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
crackmapexec==5.1.4.dev0
crcelk==1.3
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
dbus-python==1.2.16
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.10
deprecation==2.0.7
dhcpig==0.0.0
dicttoxml==1.7.4
distro==1.5.0
dj-database-url==0.5.0
Django==3.1.4
djangorestframework==3.12.2
dnslib==0.9.14
dnspython==2.0.0
EasyProcess==0.2.5
ecdsa==0.16.0
email-validator==1.1.1
entrypoints==0.3
ExifRead==2.3.2
expiringdict==1.1.4
faraday-client==1.0.0
faraday-plugins==1.2
faradaysec==3.12
feedparser==5.2.1
fierce==1.4.0
filedepot==0.5.2
filteralchemy==0.1.0
flasgger==0.9.5
Flask-BabelEx==0.9.4
Flask-Classful==0.14.1
Flask-HTTPAuth==3.2.4
Flask-KVSession-fork==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Security-Too==3.4.2
Flask-WTF==0.14.3
future==0.18.2
fuzzywuzzy==0.18.0
GDAL==3.1.4
GeoIP==1.3.2
geoip2==2.9.0
geojson==2.5.0
gevent==1.4.0
git-cola==3.8
goofile==1.6
google==2.0.2
gpg===1.14.0-unknown
graphene==2.1.7
graphene-sqlalchemy==2.1.2
graphql-core==2.2.1
graphql-relay==2.0.0
greenlet==0.4.17
gvm-tools==20.10.1
h11==0.11.0
h2==4.0.0
hashID==3.1.4
hpack==4.0.0
html2text==2020.1.16
html5lib==1.1
httplib2==0.18.1
humanize==0.0.0
hupper==1.10.2
hyperframe==6.0.0
hyperlink==19.0.0
icalendar==4.0.3
idna==2.10
imageio==2.4.1
impacket==0.9.22
importlib-metadata==1.6.0
incremental==17.5.0
intervaltree==3.0.2
invoke==1.4.1
ipdb==0.13.3
ipwhois==0.15.1
IPy==1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.7.0
kaitaistruct==0.8
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2019.9.1
KismetCaptureRtladsb==2019.10.1
KismetCaptureRtlamr==2019.10.1
kismetexternal==2019.12.1
kiwisolver==1.3.1
ldap3==2.7
ldapdomaindump==0.9.3
LibAppArmor==2.13.5
logutils==0.3.3
lsassy==2.1.2
lxml==4.6.1
lz4==3.1.1+dfsg
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.19.0
mate-menu==20.4.1
mate-tweak==21.4.1
matplotlib==3.3.2
maxminddb==1.4.1
mechanize==0.4.5
meld==3.20.2
mercurial==5.5.2
minidump==0.0.12
minikerberos==0.2.0
mistune==0.8.4
mitmproxy==5.1.1
mock==4.0.2
more-itertools==4.2.0
msgpack==1.0.0
msldap==0.2.10
multidict==5.0.0
mutagen==1.45.1
mysqlclient==1.4.4
nassl==3.1.0
nbconvert==5.6.1
nbformat==5.0.8
neo4j==1.7.0.dev0
neobolt==1.7.17
neotime==1.7.4
netaddr==0.7.19
netifaces==0.10.9
networkx==2.5
nose==1.3.7
notebook==6.1.5
nplusone==1.0.0
ntlm-auth==1.4.0
numpy==1.19.4
nyx==2.1.0
olefile==0.46
onboard==1.4.1
onioncircuits==0.7
onionshare==2.2
ospd==20.8.1
ospd-openvas==20.8.0
packaging==20.4
pandocfilters==1.4.2
paramiko==2.7.2
parsedatetime==2.4
parsero==0.82
parso==0.7.0
passlib==1.7.2
Paste==3.5.0
PasteDeploy==2.1.1
PasteScript==2.0.2
patator==0.9
pbkdf2==1.3
pbr==5.5.0
pcapy==0.11.4
pefile==2019.4.18
pexpect==4.6.0
pgspecial==1.11.10
phonenumbers==8.12.1
pickleshare==0.7.5
Pillow==8.0.1
plaster==1.0
plaster-pastedeploy==0.5
plecost==1.1.2
plotly==4.12.0
pluginbase==1.0.0
ply==3.11
prettytable==0.7.2
prometheus-client==0.7.1
promise==2.2
prompt-toolkit==3.0.8
protobuf==3.12.3
psutil==5.7.3
psycopg2==2.8.5
publicsuffix2==2.20191221
py-ubjson==0.14.0
pyaff4==0.26.post6
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.16.2
pycares==3.1.1
pycountry==20.7.3
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.9.7
pycups==2.0.1
pycurl==7.43.0.6
pydot==1.4.1
pyee==7.0.2
pyelftools==0.27
pyenchant==3.0.1
Pygments==2.7.1
PyGObject==3.38.0
pygraphviz==1.6
PyHamcrest==1.9.0
PyICU==2.5
pyinotify==0.9.6
pylnk3==0.2.1
pymongo==3.11.0
pymssql==2.1.4
PyNaCl==1.4.0
pynvim==0.4.2
pyodbc==4.0.30
PyOpenGL==3.1.5
pyOpenSSL==19.1.0
pyotp==2.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyperclip==1.8.0
pypng==0.0.20
pyppeteer==0.2.2
pypsrp==0.4.0
pypykatz==0.3.7
PyQRCode==1.2.1
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqtgraph==0.11.0
pyramid==1.10.4
PyRIC==0.1.6.4
pyrsistent==0.15.5
pyrtlsdr==0.2.91
pyserial==3.5b0
PySimpleSOAP==1.16.2
pysmbc==1.0.23
pysmi==0.3.2
pysnmp==4.4.12
PySocks==1.7.1
python-apt==2.1.5
python-dateutil==2.8.1
python-debian==0.1.38
python-debianbts==3.0.2
python-decouple==3.4
python-docx==0.8.10
python-editor==1.0.3
python-gvm==20.9.1
python-Levenshtein==0.12.0
python-libnmap==0.7.0
python-magic==0.4.16
python-magic-ahupp==0.4.13
python-memcached==1.58
python-pam==1.8.4
python-pptx==0.6.18
python-snappy==0.5.3
python-xapp==2.0.1
python-xlib==0.29
PyTrie==0.2
pytsk3==20200117
pytz==2020.4
pyusb==1.0.2
PyVirtualDisplay==0.2.1
pywerview==0.2.0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==20.0.0
qrcode==6.1
QtPy==1.9.0
rdflib==4.2.2
redis==3.3.11
rekall-core==1.7.2rc1
rekall-lib==1.7.1
reportbug==7.8.0
repoze.lru==0.7
requests==2.24.0
requests-file==1.5.1
requests-ntlm==1.1.0
requests-toolbelt==0.8.0
retrying==1.3.3
rfcat==1.0.1
roguehostapd==1.1.2
Routes==2.4.1
rq==1.5.2
rtlsdr-scanner==1.3.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
rule-engine==1.1.0
Rx==1.6.1
scapy==2.4.4
scipy==1.5.4
secure==0.2.1
selenium==4.0.0a1
Send2Trash==1.5.0
service-identity==18.1.0
setproctitle==1.2.1
shodan==1.24.0
simplejson==3.17.0
simplekv==0.13.0
sip==4.19.24
six==1.15.0
smoke-zephyr==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0.1
SPARQLWrapper==1.8.5
specan==0.0.0
SQLAlchemy==1.3.20
sqlalchemy-schemadisplay==1.3
sqlparse==0.3.1
sslyze==3.1.0
stem==1.8.0
syslog-rfc5424-formatter==1.2.2
tabulate==0.8.7
Tempita==0.5.2
termcolor==1.1.0
terminado==0.9.1
terminaltables==3.1.0
termineter==1.0.4
testpath==0.4.4
texttable==1.6.3
theHarvester==3.2.0
tld==0.11.11
tls-parser==1.2.2
torbrowser-launcher==0.3.3
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
translationstring==1.4
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
u-msgpack-python==2.3.0
ufw==0.36
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.11
urwid==2.1.1
uvloop==0.14.0
venusian==3.0.0
vinetto==0.8.0
visvis==1.11.1
wafw00f==2.1.0
wapiti3==3.0.3
wcwidth==0.1.9
webargs==6.1.0
webencodings==0.5.1
WebOb==1.8.6
websocket-client==0.57.0
websockets==8.1
websploit==4.0.4
Werkzeug==1.0.1
wfuzz==3.1.0
whois==0.8
widgetsnbextension==2.0.0
wifiphisher==1.4
wifite==2.5.7
winacl==0.0.2
wrapt==1.12.1
wsaccel==0.6.2
wsproto==0.15.0
WTForms==2.2.1
wxPython==4.0.7
xcffib==0.8.1
XlsxWriter==1.1.2
xmltodict==0.12.0
yara-python==4.0.2
yarl==1.4.2
yaswfp==0.9.3
zipp==1.0.0
zope.deprecation==4.4.0
zope.interface==5.1.2
